Feat: 닉네임 검증을 위한 NicknameValidator 추가
- 마이페이지에서 닉네임 변경하는 경우, 해당 로직 재사용성 가능성이 있어 따로 NicknameValidator 생성하였음
- Domain 폴더 생성시 옮기는 것이 좋아보임 !
