+++ Projects/Presentation/Login/Sources/Nickname/NicknameView.swift
@@ -111,12 +111,6 @@ extension NicknameView {
       textFieldStyle.updateUI(isError: error != nil)
       isNicknameCompleted = (error == nil)
     }
-    
-    private func containsSpecialCharacter(_ nickname: String) -> Bool {
-      let pattern = "^[a-zA-Z0-9가-힣]+$"
-      let regex = try! NSRegularExpression(pattern: pattern)
-      return regex.firstMatch(in: nickname, range: NSRange(location: 0, length: nickname.utf16.count)) == nil
-    }
   }
   
   // 닉네임 입력 필드 UI 관리
@@ -134,17 +128,6 @@ extension NicknameView {
     
     func updateMessage(for error: NicknameError?) {
       self.text = error?.message ?? ""
-    }
-  }
-  
-  // 경고 메시지 Enum
-  private enum NicknameError: String {
-    case tooLong = "6자 이하로 입력해주세요."
-    case invalidFormat = "영문, 숫자, 완성된 한글 조합만 가능해요."
-    case duplicate = "중복되는 닉네임입니다."
-    
-    var message: String {
-      return self.rawValue
     }
   }
